NEXUS-6277: Minimal optimization
Huge performance gain for an edge case, when accessed
repository does have needed permission, but in the same
time it's referenced by (potentially huge) group reposes
where it's member.

So, split the authc process in two, first check
perms of the given repository, and if not success,
then start to do heavy lifting by recursively gathering
all the perms of groups (and groups of groups) where
repo is member directly (or indirectly) and evaluate
the recursively collected permissions.